[Refactor] Remove obsolete enums from global/response and update imports
