+++ src/main/java/com/gdg/coffee/domain/bean/service/impl/BeanServiceImpl.java
@@ -3,13 +3,13 @@ package com.gdg.coffee.domain.bean.service.impl;
 import com.gdg.coffee.domain.bean.domain.Bean;
 import com.gdg.coffee.domain.bean.dto.BeanRequestDto;
 import com.gdg.coffee.domain.bean.dto.BeanResponseDto;
+import com.gdg.coffee.domain.bean.exception.BeanErrorCode;
 import com.gdg.coffee.domain.bean.exception.BeanException;
 import com.gdg.coffee.domain.bean.repository.BeanRepository;
 import com.gdg.coffee.domain.bean.service.BeanService;
+import com.gdg.coffee.domain.cafe.exception.CafeErrorCode;
 import com.gdg.coffee.domain.cafe.exception.CafeException;
 import com.gdg.coffee.domain.cafe.repository.CafeRepository;
-import com.gdg.coffee.global.common.response.bean.BeanErrorCode;
-import com.gdg.coffee.global.common.response.cafe.CafeErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
